% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldr_names.R
\name{ldr_names}
\alias{ldr_names}
\title{Find out which plots are included in lidar coverage}
\usage{
ldr_names(plotID, crdnt_x, crdnt_y, radius = 1)
}
\arguments{
\item{plotID}{Column with names of the plots}

\item{crdnt_x}{Column with x-coordinates (in UTM)}

\item{crdnt_y}{Column with y-coordinates (in UTM)}

\item{radius}{Search radius for Lidar points (should be chosen small to work quick)}
}
\description{
This function produces a character vector with names of plots that are
covered by lidar data
}
\author{
Alice Ziegler
}

