% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ldr_query.R
\name{ldr_query}
\alias{ldr_query}
\alias{ldr_query,addExtent}
\title{View extent/bbox of spatial objects interactively}
\usage{
ldr_query(plotID, crdnt_x, crdnt_y, radius = 1)
}
\arguments{
\item{x}{either a Raster* object or a Spatial* object}

\item{map}{a leaflet map the extent should be added to. If NULL
standard background layers are cretaed}

\item{map.types}{the map types to be used in case map is NULL}

\item{...}{additional arguments passed on to \code{\link{addRectangles}}}
}
\description{
This function produces an interactive GIS-like view of the extent/bbox
of the supplied spatial object
}
\author{
Tim Appelhans
}

