% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdc_by_ldr.R
\name{rdc_by_ldr}
\alias{rdc_by_ldr}
\title{Reduces dataframe to entries that are included in lidar coverage only}
\usage{
rdc_by_ldr(dataframe_plts, dataframe_crdnt = NULL, v_plts, ...)
}
\arguments{
\item{dataframe_plts}{Dataframe that should be reduced in the end.}

\item{dataframe_crdnt}{Dataframe with plot-ID, x-coordinates and y-coordinates
in three columns named: "plotID", "x_pnt" and "y_pnt". Default = NULL}

\item{v_plts}{Character vector with unique name entry for each plot that
should be kept during the reduction. Not necessary, if dataframe_crdnt is
provided.}

\item{...}{Argument search radius for Lidar points (should be chosen small
to work quick) can be provided.}
}
\description{
This function reduces a dataframe, so that only entries are kept that are in
a vector (by names). If no list exists yet, a dataframe with names of plots
and coordinates can be passed so that the function "ldr_names" can create
such a vector.
}
\author{
Alice Ziegler
}

