% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldr_query.R
\name{ldr_query}
\alias{ldr_query}
\title{Get statistical information about lidar data on a plot}
\usage{
ldr_query(plotID, crdnt_x, crdnt_y, radius, height = F)
}
\arguments{
\item{dataframe_crdnt}{Dataframe with plot-ID, x-coordinates and y-coordinates
in three columns named: "plotID", "x_pnt" and "y_pnt".}
}
\description{
This function produces a dataframe with statistical information about lidar data
}
\author{
Alice Ziegler
}

